Move configurator debug data into configurator.
The current chaos_base_test knows how to run the connection test, and it
knows how to debug APConfigurator.  As we move to the model of having an
APRunner and the tests looking more like the wifi cell tests we need to
move this logic somewhere else.  It makes the most sense to keep this info
in the APConfigurator itself.

BUG=None
TEST=Unittest, ran network_WiFiChaosOpen/control.debug$

Change-Id: I0056089a6524bb794c635bb1723257170489a0ba
Reviewed-on: https://chromium-review.googlesource.com/168881
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>