Add network_WlanDriver to a couple of suites
Add it to network_nightly in order to continuously test all boards
in the scheduler.  Add it to wifi_matfunc so that manual tests of
boards that do not live in our test cells full-time can be covered.

BUG=chromium:289778
TEST=Ran this test against all platforms in
http://cautotest/afe/#tab_id=view_job&object_id=4743561.  Only
failing units had old images on them.

Change-Id: I35281f3003240426db1dc19584c068fdb0368241
Reviewed-on: https://chromium-review.googlesource.com/169445
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>