[Autotest] Add default host parameters to host factory.
There are times when we would like to create a host outside the context of
a server job. One such scenario is with test_that, where we would like to ssh
into a DUT and determine its board if one wasn't specified. Today this is
impossible without hacking in the values of ssh_user, ssh_pass etc, directly
into the factory, or incurring the overhead of an extra test run that will
drop the board into a keyval.

This cl uses the same default arguments the server_job uses if ssh_user, ssh_pass
etc are not specifiec thus enabling us to create hosts on-the-fly.

TEST=Created hosts, used them to get board from test_that. Ran run_suite
     with a dummy suite to confirm that server_job can still create hosts.
BUG=chromium:265705

Change-Id: I50811eb8b0e6b145a190211391c2409913bc8a33
Reviewed-on: https://chromium-review.googlesource.com/169727
Commit-Queue: Prashanth B <beeps@chromium.org>
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Prashanth B <beeps@chromium.org>