[autotest] Add a provision test for factory images.
This test enables us to reimage a servo connected DUT to a factory
image through test_that. It makes use of functionality added in previous
cls to understand that a DUT has a factory image, and uses a more relaxed
usb boot timeout when performing the reimage.

TEST=Reimaged DUTs.
BUG=chromium:279470

Change-Id: I49fb4ea18de63b8579beea3a6bfdb50ea9915cee
Reviewed-on: https://chromium-review.googlesource.com/169939
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>