lansim: Add IP and UDPSocket support to the Host interface.
This patch adds a new "socket()" call to the SimpleHost class that
provides an asynchronous socket like interface. These interface
allows the caller to send and receive UDP sockets as in a normal
socket interface.

This intermediate socket interface makes it easier to replace the
lansim backend with a different backed based on real kernel sockets
to allow reuse the implementations above it on different
environments.

BUG=chromium:288385
TEST=Manual test using Zeroconf and P2P.

Change-Id: I3866aab38e581927a424a2e747ce208dc413c0b9
Reviewed-on: https://chromium-review.googlesource.com/171692
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>