[autotest] test_that respects a test's DEPENDENCIES.
test_that now auto-detects a host's labels and assign appropriate labels
to the DUT in afe. When the suite is scheduled locally, only test
with satisfiable DEPENDIENCIES will be scheduled.

There is no need to check with each job's dependencies.
Because only one DUT is added to the AFE.
If no host(in our case, only one host) can statisfy a
test's dependencies, Suite.schedule() will not create
a job for that test.

BUG=chromium:265715
TEST=Run dummy suite, confirm dummy_BT is not run because it
requests bluetooth; Once manually add 'bluetooth' label to
the dut, dummy_BT is scheduled and run. Run dummy_BT,
confirm it is not run.

Change-Id: Icc653f79b69ea87247082662306b6af64b1b970b
Reviewed-on: https://chromium-review.googlesource.com/171444
Reviewed-by: Fang Deng <fdeng@chromium.org>
Commit-Queue: Fang Deng <fdeng@chromium.org>
Tested-by: Fang Deng <fdeng@chromium.org>