power_LoadTest: Initialize _services to None
When there's a failure in the initialize function or the beginning of run_once,
an exception will be thrown in the cleanup function when we try to access the
_services member. This is because the _services member has not been added to the
class instance yet, so an unhandled exception is thrown. Initializing _services
to None fixes this. This will help debug whatever issue was causing initialize
or run_once to fail.

BUG=chromium:297830
TEST=run power_LoadTest (just check that it starts running)

Change-Id: I1e93ba4a6e9681a30464209bc34e2251253762e0
Reviewed-on: https://chromium-review.googlesource.com/170530
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>