Allow run_remote_tests to proceed with inaccessible sheriffs
A modification in CL169724 made the run_remote_tests script
retrieve sheriffs for the google build tree for reporting
failures.  Outside of google, behind a typical proxied
corporate firewall, the defauilt URL specified in the
global_config.ini file fails the urlopen().read() call with
a httplib.BadStatusLine exception.  This was NOT caught by
the default handler that logs a message and allows the
caller to proceed without that sheriff.  This regression
caused the entire run_remote_tests script to fail.

This patch adds an additional exception handler to catch a
wider range of exceptions, log an message, and proceed
without that sheriff.

BUG=none
TEST=run_remote_tests from behind proxied firewall

Change-Id: Ife65b244291acd91117553f0a68ef661f267040d
Reviewed-on: https://chromium-review.googlesource.com/171481
Tested-by: Paul Drews <paul.drews@intel.com>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Paul Drews <paul.drews@intel.com>