Autotest: Adds wifi_noise_level method to wifi_client.
This new method returns the noise level for a specified frequency on
the DUT.  Currently, this method only returns valid values for versions
of ChromeOS with drivers that report the value.  On other devices, this
method returns None.

While I was there (and due to an excellent code review comment), I
did a tiny refactor to keep from repeating the 'iw' command that checks
a device to see if it's a wifi device.

BUG=None
TEST=manual:

* add a line in an autotest that looks like:
    log.logging('NOISE: %r',
                self.context.client.wifi_noise_level(frequency))
  'frequency' needs to be a valid WiFi frequency.
* on the DUT in a separate window, type 'ifconfig -a'
* Then, using the wifi interface reported by ifconfig, type:
    'iw dev <wifi-xface> survey dump'
* Verify that the noise value under the frequency used in the autotest
  is about the same (it will vary a little from call to call) as that
  reported by the autotest.

Change-Id: I028cd1d0440c6a3ea67245856d0a80b84c9a6223
Reviewed-on: https://chromium-review.googlesource.com/170682
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Commit-Queue: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>