autotest: Adds 'clear_supplicant_blacklist'
wpa_supplicant adds a BSSID to its blacklist (i.e., it won't roam to the
BSS) if supplicant gets a DEAUTH for the BSSID.  This happens, for
example, when if (like most of our tests) a router is deconfigured
without the currently-connected ssid having been disconnected.  For
single radio tests, this isn't a big deal: supplicant clears the
blacklist if a scan finds no available non-blacklisted BSSes.  For
two radio tests, though, this leaves one BSSID blacklisted and unusable.
This provides an API to clear the blacklist.

BUG=None
TEST=autotest and maunal.  For the manual test, do the following:

For all work on the DUT, below, you can either
  * ssh to the DUT
or:
  * use a crosh tab on the dut (ctrl-alt-T)
  * > shell
  * > sudo bash

Verify that things work OK with when clearing the blacklist:

  * On the DUT:
    * > echo '' > /var/log/net.log

  * On the workstation (in a chroot):
    * > test_that --fast -b ${BOARD} chromeos1-grover-host1.cros \
        network_WiFi_SSIDMultiSwitchBackLowPower
    * Verify that the test passes.

  * On the DUT:
    * > grep "Removed BSSID .* from blacklist (clear)" /var/log/net.log
    * Verify that at least one entry appears with a MAC address that is
      not 00:00:00:00:00:00

Verify that things do not work when not clearing the blacklist:

  * On the workstation:
    * In this file:
        src/third_party/autotest/files/server/site_tests/
        network_WiFi_SSIDMultiSwitchBackLowPower/
        network_WiFi_SSIDMultiSwitchBackLowPower.py

      Comment out the following line:
        self.context.client.clear_supplicant_blacklist()

  * On the DUT:
    * > echo '' > /var/log/net.log

  * On the workstation (in a chroot):
    * > test_that --fast -b ${BOARD} chromeos1-grover-host1.cros \
        network_WiFi_SSIDMultiSwitchBackLowPower
    * If the test passes the first time (and it might), re-clear
      /var/log/net.log (as above) and rerun the test.
    * Verify that the test fails.

  * On the DUT:
    * > grep "Removed BSSID .* from blacklist (clear)" /var/log/net.log
    * Verify that at no entry appears with a MAC address that is
      not 00:00:00:00:00:00

CQ-DEPEND=I1274c3fed10e93386ffc49e7a3e0b45c3e98856b

Change-Id: Id633bb69ad8abb82f207c27043c2ddbbdddaf332
Reviewed-on: https://chromium-review.googlesource.com/170696
Tested-by: Wade Guthrie <wdg@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Commit-Queue: Wade Guthrie <wdg@chromium.org>