[autotest] Deduping scheduler cannot file bugs with owner=sheriff.
Sometimes the sheriff isn't available because the waterfall is managed
by the channel. The bug filer is unforgiving when we try to assign the
bug to an invalid id (eg:foo@chromium.org), but will silently move on
if we try to cc an invalid id.

This cl cc's the sheriff to deduping_scheduler bugs, and adds the Hardware-lab
label so these bugs don't go unnoticed when the waterfall is managed by the
channel. It also fixes the error message in the bug filer to surface bad owner
failures.

TEST=Filed bugs with an invalid sheriff as cc.
BUG=chromium:301140

Change-Id: I6a588cd0601cb9d4e62d0bcf38d336766341bddf
Reviewed-on: https://chromium-review.googlesource.com/171008
Reviewed-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>
Tested-by: Prashanth B <beeps@chromium.org>