Autotest: Autodetect ADB Host.
Allow tests to run on either ADBHost or CrosHost without requiring
the host type to be specified. If the 'adb' command exists on the
device it will use ADBHost otherwise it will use CrosHost.

Measured the cost of creating an SSHHost instance and running
'which adb' on editted control file in the AFE and it added roughly
500 milliseconds to the test completion time.

BUG=chromium:294328
TEST=Ran locally and tested with a try-bot to verify VMTests still
work.

Change-Id: I0ca2a15e8473284e4e52c3176b933a136b341d79
Reviewed-on: https://chromium-review.googlesource.com/171092
Reviewed-by: Fang Deng <fdeng@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>