Temporarily turn-off unittests in wardmodem till all CLs are in.
There are a bunch of wardmodem CLs that are ready to be merged. Turning off the
unit-tests as they go in will simplify the process.
Wardmodem is not used by any tests right now.
This CL will be reverted once these CLs are in.

BUG=chromium:302840
TEST=Run unittest_suite.py, ensure new tests from wardmodem are not run.

Change-Id: Ib951428781880044f9fb0935764fcbf1b7b5f76f
Reviewed-on: https://chromium-review.googlesource.com/171500
Reviewed-by: Arman Uguray <armansito@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>