autotest: Fix site_wifitest errors for tests in wifi_matfunc suite
site_wifitest was erroneously trying to call sync_host_times on the host
objects for the client, router, and server rather than the appropriate
wrappers.  The command lookup was broken on RSPro/Netgear setups.  RSPro
cells need a different command line to start hostapd and send its logs
to a file.  The hostapd on those setups also doesn't seem to write out
its PID file, so move the logic that checks hostapd start health

BUG=None
TEST=network_WiFiManager tests pass again on stumpy cells.  092* passes
on an RSPro cell.  The remaining 3 Manager tests still fail on RSPro
cells, but this is because wait_service expects to find that the service
is online, when it is only portalled in those setups.

Change-Id: I5db662d1567bed812c7a1879b750d870bbed0048
Reviewed-on: https://chromium-review.googlesource.com/171792
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>