lansim: Add TCP streams forwarding.
This patch includes support to forward incoming TCP connections on
any simulated SimpleHost back to a real host IP address on the same
network. This lets the code under test make a connection to a fake
TCP server running on a virtual host be redirected to a real host
and use the kernel's network stack to handle it.

BUG=chromium:241519
TEST=Added new unittests to lansim.

Change-Id: Id0700ef4a9c78ac17e183c20e33f5dfe8550c4f9
Reviewed-on: https://chromium-review.googlesource.com/171817
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>