Fix firmware_ECPowerG3 test.
With new haswell systems, we cannot detect the intermediate S5 state before
they go into G3. So removing that check which consistently failed on haswell
systems.

Note that this test change impacts older platforms as well, and is not
limitted to only to haswell systems.
So any assumption that we verify that older systems go into S5 before G3 will
not be true.

Also renamed "timeout" to "retries" which reflect the actual use.

BUG=chrome-os-partner:22021
TEST=Manual tests at desk with peppy device

Change-Id: I0ff181dc9ca266d5889152a06a620c01908660a5
Reviewed-on: https://chromium-review.googlesource.com/172021
Tested-by: Yusuf Mohsinally <mohsinally@chromium.org>
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Commit-Queue: Yusuf Mohsinally <mohsinally@chromium.org>