Repeat bluetooth_Sanity_Discovery/Discoverable
Since these tests involve radio functions, they are not 100% reliable on
any given run. Repeat the tests for up to five times in the case of
failure, which should produce a positive result.

Use perf keyvals both to record the number of failed attempts (hopefully
we'll find a way to bring this to zero and remove the repetition) and
also record the signal strength (another useful value).

BUG=chromium:256771
TEST=test_that bluetooth_Sanity_Discovery;
 test_that bluetooth_Sanity_Discoverable

Change-Id: I335e2fdc39c695c2ae6ecdc2dba4b5b73c43526d
Reviewed-on: https://chromium-review.googlesource.com/172170
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Scott James Remnant <keybuk@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>