autotest: Refine WpaCliProxy connect_wifi() logic
Internally, wpa_supplicant builds its scan_results reply in a 4K byte
buffer.  However, with lots of APs, we can run out of space in the
buffer and not see our intended BSS.  Tolerate this by short-circuiting
out of the discovery wait period if we discover that we've already
associated.

Also, not all routers happen to put us on a 192.168.0.0/16 subnet.
Don't hardcode that expectation.

BUG=chromium:305801
TEST=Had a successful interop run with these changes.

Change-Id: Ib85504fc0f521d79cef24342b821b7211841fd3e
Reviewed-on: https://chromium-review.googlesource.com/172502
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>