Move login_status_ext to autotest_private_ext.
A number of tests need access to autotestPrivate-
now chrome.py provides access to an extension with this api by simply setting a boolean.
A number of tests could use login_status as well, so add a property for this.
Add support for oobe access.
Refactor telemetry_LoginTest to use the new properties.

BUG=chromium:233834
TEST=telemetry_LoginTest

Change-Id: Ia6474ff9b280a7bbda2ebeb9edbdc8c20169cfde
Reviewed-on: https://chromium-review.googlesource.com/172507
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Commit-Queue: Achuith Bhandarkar <achuith@chromium.org>
Tested-by: Achuith Bhandarkar <achuith@chromium.org>