client: log exceptions generated while building client tests.
When a client test is imported in the chroot, it might fail to import
for various reasons. The current code only shows the test name and
the exception message, but not the traceback, making it difficult to
track why it fails to import on the autotest context.

This patch also logs the exception traceback in this case.

BUG=None
TEST=emerge-daisy_spring autotest-tests shows why and where my test is failing to import.

Change-Id: I8f41f308790c410ab49ac29cba9c63ba31419f06
Reviewed-on: https://chromium-review.googlesource.com/172623
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>