p2p: New p2p_ShareFiles test.
This test passively checks that the DUT exposes the right information
over mDNS when the service p2p is running. To achieve this a new
CrosP2PClient class implements the client part of the P2P protocol,
but only in a passive way. This means that this fake client doesn't
send any request to the DUT. Nevertheless, the mDNS/DNS-SD protocol
states that the DUT should annouce the records to the multicast
address. This can be extended later to actively request the same
information.

BUG=chromium:271847
TEST=test_that --fast -b link <dut_ip> p2p_ShareFiles

Change-Id: I9f4e61e3094dd876dfaf749df051d42ebf8f59a8
Reviewed-on: https://chromium-review.googlesource.com/172894
Reviewed-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>