Autotest: Fix server-side profilers and support telemetry_Crosperf
Server-side profiler support was written back in 2010 and there does
not seem to be any recent usages of it thus it no longer worked.
Discovered this is due to chrome-os not allowing incoming connections
on the port the barrier's listening server uses.

This change adds the command to open the port and make server side
profilers work as expected. It also fixes us the arg parsing of
telemetry_Crosperf so the profiler runs properly.

BUG=chromium:303925
TEST=test_that --board=link --args="iterations=1 test=octane
profiler=custom_perf profiler_args='record -a -a -e cycles,instructions'"
172.22.75.140 telemetry_Crosperf

Change-Id: I35952faaaa7aa829653898f852216b4ffe3d6342
Reviewed-on: https://chromium-review.googlesource.com/173086
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>