Refactor verification of booted partition logic so it's reusable.
The verification logic doesn't really belong to cros_host.py and putting
it there makes it difficult to re-use across server-side tests. In
a follow-up CL I plan to use verify_kernel_boot alone (to verify we
switched partitions in an update the test controls). This CL makes it
easy to do that.

BUG=chromium:260468
TEST=Ran with new test, pylint, pyflakes.

Change-Id: If81a65b3cb388e8ac5a55dbb83ba2e1617d326a2
Reviewed-on: https://chromium-review.googlesource.com/173219
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>