autotest: Clean up SSH tunnels before starting XMLRPC proxies
In FAFT tests, they restart the DUT multiple times in a test, which
requires the XMLRPC server to be restarted each time.  However, the
tunnel to the previous XMLRPC server still exists for all but the first
attempt.  For each of these attempts, we:

1) Start an XMLRPC server
2) Clean up existing SSH tunnels
3) Open a new SSH tunnel

Unfortunately, part of cleaning up existing SSH tunnels is terminating
server processes at the end of those tunnels.  This kills the server
we had just started in step 1).

Fix this by changing the sequence to look like:

1) Clean up existing SSH tunnels
2) Start an XMLRPC server
3) Clean up existing SSH tunnels (noop)
4) Open a new SSH tunnel

TEST=firmware_FAFTSetup passes.
BUG=chromium:308198

Change-Id: Ie387f69005759de04b488d31567746cee764744c
Reviewed-on: https://chromium-review.googlesource.com/173236
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Yusuf Mohsinally <mohsinally@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>