Autoserv: Log SIGTERM Signals.
To help debug what is happening with the suites dying without writing
exit code, lets log when SIGTERM is received by autoserv. If this
message appears missing when the errors occur, the process would not
have received SIGTERM and if it does receive the signal it disproves
the theory the signal is not received.

BUG=chromium:302815
TEST=local setup

Change-Id: Ie6a72da1e1758998825d292f5a0e577eb6e90572
Reviewed-on: https://chromium-review.googlesource.com/173325
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>