[autotest] fix a race condition in master-ssh startup
This fixes a race condition in master-ssh connection startup, between
the establishment of the connection (and creation of socket file) and
the first use of the connection. It accomplishes the fix by waiting up
to 5s for the socket file to appear, after launching the master-ssh
process, before returning from the master-ssh start function. If the
connection fails to be established within this time, a warning is logged
but no other action is taken, allowing the test to attempt to continue
to run.

BUG=chromium:305641
TEST=Ran some local test_that runs, which prior to this change had been
exhibiting a lot of "lost socket file" events in the logs. Saw these
events eliminted.

Change-Id: Ie8a45edf20adf1320a65201451ab7b64ac837dbb
Reviewed-on: https://chromium-review.googlesource.com/173560
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>