Initial add of a touchbotII robot_wrapper.py
Right now this module doesn't get used by anything as I want to maintain
the functionality of the FW test with the old robot until this one is
fully functional.  As such, I've made a separate module that can easily
be swapped in for testing with just a couple manual changes.  This way I
can work on this module next to the functioning one and once it's fully
in shape I can swap them permanently.

This is based heavily off robot_wrapper.py but I wanted to get eyes on
it before I went too far so this wouldn't turn into a monster of a code
review.

Essentially this is a robot wrapper that only supports 1&2f tracking
gestures and "rapid taps" by using the drumroll gesture and the line
gesture that have already been added to the touchbot control module.  I
will be adding to this as I go filling in the gaps, until hopefully all
the current fw tests are included and more after that.

BUG=chromium:308162
TEST=run the robot test on a machine and it performed the enabled
gestures admirably when I swapped this file out for the existing
robot_wrapper and disabled the unsupported tests.

Change-Id: I417e03f55434b1000a9bb71ea8bfa17f9588ea95
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/173596
Commit-Queue: David James <davidjames@chromium.org>