[autotest] Force to collect all logs if test failed.
The default behavior is to save only new data in /var/log. This CL changes that
behavior so that, if a test failed, all files in /var/log will be collected.

BUG=chromium:303352
TEST=test_that 172.22.75.225 dummy_Fail
check results folder that contains all files in /var/log:
/tmp/test_that_results_0JCjJZ/results-1-dummy_Fail/dummy_Fail.Fail/sysinfo/var/log/
If the test passed, the folder should be var/log_diff.

Change-Id: I4c876c4a25a129841ff82763412aec2b317825a1
Reviewed-on: https://chromium-review.googlesource.com/174058
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>