Autotest: Add logging for drones.
Currently we have no way to log what occurs on our drone machines.
They can send warnings back to the scheduler and thats about it.

This change creates a daily drone log file so that we can add and
understand better how the drone interacts with it's autoserv processes.

NOTE: This change is implemented in the case the Drone is running
remotely and has its main method executed. IF it is a local drone,
the logging messages will simply fold into the Scheduler's logs as
in this situation, drone_utility is called directly.

BUG=None
TEST=set my shadow config to point Drones at my machine (rather than
localhost) and ensured the new log file appeared and has my log
messages.

Change-Id: I6824d4a60d6190f40e888ba7bf952c5d704fae3b
Reviewed-on: https://chromium-review.googlesource.com/174372
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>