Add Touchbot II support for swipes in FW test
This adds the plumbing into the Touch FW test to interface with
the Touchbot II and it's "swipe" gesture.  Essentially this is just a
small modification of the line parsing to add an additional parameter to
the end.  This also allows the basic lines to work with the newly
modified line.py gesture since it takes an additional parameter now that
the old version of this wasn't supplying.

BUG=chromium:308162
TEST=Manually ran the FW test with the robot and it parsed the gestures
and variations just fine and performed them on the pad.

Change-Id: I231f3631faff84cd6ca51094c7a70afe2c44afb6
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/174420
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>