[autotest] Machine death reporting now uses bug filer.
The machine death reporting code used to log incidents to a special file.
Now it uses the bug reporter. This will submit a bug for each failed repair
attempt. If bug reporting fails, this modules will send out an email,
and if sending an email fails it will revert to the old behavior of just
logging the fact.

BUG=chromium:260031
TEST=Unittests, filed bugs, sent mail, checked logs.

Change-Id: I3ac36efb8a9e0c4d2665e9a2d42f64557e2323a0
Reviewed-on: https://chromium-review.googlesource.com/174622
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>