p2p_ShareFiles: Reduce flakiness waiting for events instead of timeouts.
This patch adds a new "until" argument to the lansim Simulator class in
order to run the simulator *until* a condition is met. This allows the
p2p_ShareFiles to wait *until* the required condition is met, and set
a higher timeout for failure. This should reduce the flakiness of the
test and improve the running time.

BUG=None
TEST=New unittest passes. test_that p2p_ShareFiles still passes.

Change-Id: I086546bcada018c24f10bcffd51127260dbc4c30
Reviewed-on: https://chromium-review.googlesource.com/174906
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>