Fixes, better messages in TestError and TestFail
There is a lot of try code when we attempt to grab the screenshot. I
found a couple bugs there as that code is not exercised very often.

In addition, I added more information when tests Error or Fail. This
will help in bug filing.  Currently we have to use the info in the
tag to lookup the rest of the info in the chaos_dynamic_config.

Finally, we were not setting the traceback in the configurator when
and exception was thrown.  This CL fixes that.

TEST=Unittests, control.debug
BUG=chromium:312537,chromium:312539

Change-Id: I891a32e2634550c8937eee6c7af50942dbcccfd6
Reviewed-on: https://chromium-review.googlesource.com/174935
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Bindu Mahadev <bmahadev@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>