[autotest] Fix parent/child and multiple aborts.
The queue_entry_id is a primary key in afe_aborted_host_queue_entries, which
should precludes us from trying to abort the same job multiple times. It will
happen if we abort a parent job (which leads to the abortion of its children),
and then abort a child in that set before the scheduler sets the complete bit.

The complete bit is set when we abort the hqe, so depending on where the
scheduler is in the tick, this leaves a ~30 second window within which our
second child abort will lead to an IntegrityError. This window is actually much
longer if the HQE is already in PARSING, since we won't set the complete bit
till the epilog of the PostJobTask.

TEST=Aborted suites, then their children. Tried to abort complete jobs.
BUG=chromium:308010
DEPLOY=Apache

Change-Id: Ib112502e966f5575c523ac053e94d44bd92359d6
Reviewed-on: https://chromium-review.googlesource.com/174988
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>