deps: rewrite fakegudev to properly subclass GUdevDevice
FakeGUdev fakes out a GUdevDevice for client code. Before this CL, the object
returned by FakeGUdev was not a GObject object compatible with GUdevDevice. This
means that any client code that checked the type of this object would fail.

This CL rewrite FakeGUdevDevice to be a subclass of GUdevDevice, thus more
throroughly faking the object.

BUG=chromium:297876
TEST=Run test-gulib.sh. Ensure there are no crashes.

Change-Id: Ie17182684b8af52781833f82b5ae73f9c3954e1c
Reviewed-on: https://chromium-review.googlesource.com/175199
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>