use boot_id in desktopui_CrashyRebootServer
When testing to see if the DUT has gone down in response to
some expected test behavior, it's possible for the harness
to 'miss' this event if it checks just before and just after
a successful reboot. Using a 'boot_id' should allow us
to see that the device has rebooted between ssh-pings.

BUG=None
TEST=test_that ... e:desktopui_CrashyRebootServer

Change-Id: I95cae3c39df94c8595cb732c589933906d6173c7
Reviewed-on: https://chromium-review.googlesource.com/175026
Tested-by: Chris Masone <cmasone@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>