[autotest] De-prioritize hostless hqes in favor of tests.
Currently, hostless hqes get precedence over tests. In situations
when we're flooded with suites this is a problem, as it leads to
a deadlock situation where many hostless jobs are waiting on tests
that the drone doesn't have the capacity to run. Note that even
after this change such scenarios are possible, just a little less
likely.

TEST=Started suites, set a low limit, checked that we run tests before
     hostless jobs. Checked that we prioritize (host+no metahost) over
     (no host+metahost). Ran unittests.
BUG=chromium:312847
DEPLOY=scheduler

Change-Id: Ibe66e8a0b6319561cc24e491ec7b9b370a840bad
Reviewed-on: https://chromium-review.googlesource.com/175028
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>