[autotest] Add support for sharding by pool to run_suite.
This adds a POOL_INSTANCE_SHARDING variable to the config that controls
what instance it hit for a certain pool.  This is then used to submit
all RPCs to, and is used as the host to link to for results.

BUG=chromium:312918
TEST=ran cq/not cq, saw run_suite hit right host.  Correct links were
printed to get to results.

Change-Id: I1c9fd4df784467d52d8472dae0b1f172ab6ca061
Reviewed-on: https://chromium-review.googlesource.com/175163
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>