p2p_ConsumeFiles: Raise error excetions and fix test.
The test p2p_ConsumeFiles checks, among other things, that the
p2p-client client hangs when there are too many connections on the
network. This check was wrong and wasn't making the test fail because
the error.TestFail exception was being created but not raised.

This patch solves that problem and also fixes the test in a more
robust way. The fake p2p server running on the simulator can now
optionally update the network with the new shared file or number of
connections, as the real p2p-server does. This notifies p2p-clients
with a cached value of num_connections that a new value of
num_connections is available.

To achieve this, some locking between the test thread and the
simulator thread is added to allow the test thread call functions
that send packets but from the simulator thread.

BUG=None
TEST=lansim unittests, test_that p2p_ConsumeFiles and p2p_ShareFiles pass.

Change-Id: I89ebf5a33d55cb3ab0f06283f79c6c335d162368
Reviewed-on: https://chromium-review.googlesource.com/175169
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>