autotest: Move iw runner into client library
We may find it useful to use iw related logic in client tests.  To
enable this, make the host parameter optional and run the commands on
localhost if no remote host is specified.

On the way, fix a missing error import in chaos_runner.py.

TEST=Interop debug test with 2 APs still works.
BUG=None

Change-Id: I41289fc8390873c73ac3474d37f4c89a0bc6368f
Reviewed-on: https://chromium-review.googlesource.com/175366
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>