Turn up timeout in desktopui_CrashyRebootServer
In this test, we purposefully crash Chrome over and over in order to
ensure that we'll eventually reboot the device to try to mitigate a
potential GPU hang. The difficulty is detecting that the device has
rebooted without waiting forever in the case that the device never
shuts down.

The current timeout of 30s does not seem to be enough time to allow slow devices
to recover from all the UI crashes, trigger a reboot, actually shut down, and
then come back up. Since the timeout is only hit when we actually fail to shut
down at all, it's OK for it to be long. We have other tests that should allow us
to catch and fix issues that cause the device to take a really long time to
shut down.

BUG=None
TEST=desktopui_CrashyRebootServer

Change-Id: I1e0c41f360a334d0c300e9f251daa9a02449b535
Reviewed-on: https://chromium-review.googlesource.com/175555
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>