Autotest: gs_offloader add delete-only and days-old args.
Since we now have the CQ Scheduler and database, we now have another
source of results. However we don't need these results after a week or
so. So instead we modified gs_offloader to have the option to delete
only and specify the number of days to wait before deleting an results
folder.

BUG=chromium:314764
TEST=locally ran and ensured that jobs/special task results are only
deleted when requested and that the days-old argument filters out the
desired folders to delete.

Change-Id: I6696a7abd81e826737514a8f12cc46c80962121e
Reviewed-on: https://chromium-review.googlesource.com/175651
Reviewed-by: Alex Miller <milleral@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Simran Basi <sbasi@chromium.org>